crypto-symmetric/requirements.txt
pwntools=4.14.1
pycryptodome=3.23.0
